na Samsunga

+++ MobileApp1/MobileApp1/OptionsPage.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MobileApp1.Model;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -19,34 +20,90 @@ namespace MobileApp1
 
         public async void DeleteAllClicked(object sender, EventArgs e)
         {
-            bool answer = await DisplayAlert("Ostrzeżenie", "Czy na pewno chcesz usunąc wszystkie pozycje?", "Tak", "Nie");
+            var AllItems = await App.Database.GetItemsAsync();
 
-            if (answer == true)
+            if (AllItems.Count == 0)
+                await DisplayAlert("Błąd", "Brak pozycji do usunięcia.", "OK");
+
+
+            else
             {
-                var AllItems = await App.Database.GetItemsAsync();
+                bool answer = await DisplayAlert("Ostrzeżenie", "Czy na pewno chcesz usunąc wszystkie pozycje?", "Tak", "Nie");
 
-                foreach (var item in AllItems)
-                    await App.Database.DeleteItemAsync(item);
+                if (answer == true)
+                {
 
-                Application.Current.MainPage = new NavigationPage(new MainPage());
+                    foreach (var item in AllItems)
+                        await App.Database.DeleteItemAsync(item);
+
+                    Application.Current.MainPage = new NavigationPage(new MainPage());
+                }
             }
         }
         
         public async void UncheckAllClicked(object sender, EventArgs e)
         {
-            bool answer = await DisplayAlert("Ostrzeżenie", "Czy na pewno chcesz odznaczyc wszystkie pozycje?", "Tak", "Nie");
+            var AllItems = await App.Database.GetItemsAsync();
 
-            if (answer == true)
+            List<Item> items = new List<Item>();
+
+            foreach (var item in AllItems)
             {
-                var AllItems = await App.Database.GetItemsAsync();
+                if (item.IsChecked == true)
+                    items.Add(item);
+            }
 
-                foreach (var item in AllItems)
+            if (items.Count == 0)
+                await DisplayAlert("Błąd", "Brak pozycji do odznaczenia.", "OK");
+
+            else
+            {
+                bool answer = await DisplayAlert("Ostrzeżenie", "Czy na pewno chcesz odznaczyc wszystkie pozycje?", "Tak", "Nie");
+
+                if (answer == true)
                 {
-                    item.IsChecked = false;
-                    await App.Database.UpdateItemAsync(item);
-                }
 
-                Application.Current.MainPage = new NavigationPage(new MainPage());
+                    foreach (var item in AllItems)
+                    {
+                        item.IsChecked = false;
+                        await App.Database.UpdateItemAsync(item);
+                    }
+
+                    Application.Current.MainPage = new NavigationPage(new MainPage());
+                }
+            }
+        }
+        public async void DeleteCheckedClicked(object sender, EventArgs e)
+        {
+            var AllItems = await App.Database.GetItemsAsync();
+
+            List<Item> items = new List<Item>();
+
+            foreach (var item in AllItems)
+            {
+                if (item.IsChecked == true)
+                    items.Add(item);
+            }
+
+            if (items.Count == 0)
+                await DisplayAlert("Błąd", "Brak pozycji do usunięcia.", "OK");
+
+
+            else
+            {
+                bool answer = await DisplayAlert("Ostrzeżenie", "Czy na pewno chcesz usunąć kupione pozycje?", "Tak", "Nie");
+
+                if (answer == true)
+                {
+
+                    foreach (var item in AllItems)
+                    {
+                        if (item.IsChecked == true)
+                            await App.Database.DeleteItemAsync(item);
+                    }
+
+                    Application.Current.MainPage = new NavigationPage(new MainPage());
+                }
             }
         }
     }
